+++ elastic-job-cloud/elastic-job-cloud-scheduler/src/main/resources/console/js/history/job_exec_details.js
@@ -44,7 +44,8 @@ function splitFormatter(value) {
     var replacement = "...";
     if(null != value && value.length > maxLength) {
         var vauleDetail = value.substring(0 , maxLength - replacement.length) + replacement;
-        return '<a href="javascript: void(0);" style="color:#FF0000;" onClick="showHistoryMessage(\'' + value.replace(/\n/g,"<br/>") + '\')">' + vauleDetail + '</a>';
+        value = value.replace(/\r\n/g,"<br/>").replace(/\n/g,"<br/>").replace(/\'/g, "\\'");
+        return '<a href="javascript: void(0);" style="color:#FF0000;" onClick="showHistoryMessage(\'' + value + '\')">' + vauleDetail + '</a>';
     }
     return value;
 }