+++ RELEASE-NOTES.md
@@ -2,6 +2,8 @@
 
 ### 功能提升
 
+1. [ISSUE #352](https://github.com/dangdangdotcom/elastic-job/issues/352) Cloud作业本地运行
+
 ### 缺陷修正
 
 1. [ISSUE #322](https://github.com/dangdangdotcom/elastic-job/issues/322) elastic-job-cloud-scheduler调度任务评估资源时考虑对executor的资源使用情况